Move menu item in wp-story-editor
